pyproject.toml
[tool.poetry]
name = "django-project-template"
version = "0.1.0"
description = ""
authors = ["Raúl Negrón <raul.esteban.negron@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "<3.0"
djangorestframework = "^3.11.0"
django-model-utils = "^4.0.0"
django-cors-headers = "^3.2.1"
django-extensions = "^2.2.9"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"
django-filter = "^2.2.0"
dj-rest-auth = "^1.0.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
django-debug-toolbar = "^2.2"
isort = "^4.3.21"
pre-commit = "^2.2.0"
pytest-django = "^3.9.0"
factory-boy = "^2.12.0"
flake8 = "^3.7.9"

[tool.black]
line-length = 88
target-version = ['py38']
exclude = '(\.eggs|\.git|\.mypy_cache|\.venv|_build|build|dist|.*\/migrations\/.*)'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
